backend/requirements.txt
Django>=5.2,<6.0
djangorestframework>=3.16,<4.0
djangorestframework-simplejwt>=5.5,<6.0
psycopg[binary]>=3.2,<4.0
redis>=7.0,<8.0
black>=26.0,<27.0
ruff>=0.15,<1.0
google-auth>=2.40,<3.0
requests>=2.32,<3.0
